ch04/pyproject.toml
[project]
name = "ch04"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.30.1",
    "jupyter>=1.1.1",
    "jupyterlab>=4.4.6",
    "langchain-community>=0.3.29",
    "langchain-google-genai>=2.1.10",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
]
